Checkpoint as of first successful tp connection, before posix, & before mckusick starts incorporating gnodes
SCCS-vsn: sys/netiso/argo_debug.h 7.2
SCCS-vsn: sys/netiso/clnl.h 7.2
SCCS-vsn: sys/netiso/clnp.h 7.4
SCCS-vsn: sys/netiso/clnp_debug.c 7.2
SCCS-vsn: sys/netiso/clnp_er.c 7.3
SCCS-vsn: sys/netiso/clnp_frag.c 7.3
SCCS-vsn: sys/netiso/clnp_input.c 7.4
SCCS-vsn: sys/netiso/clnp_options.c 7.3
SCCS-vsn: sys/netiso/clnp_output.c 7.3
SCCS-vsn: sys/netiso/clnp_raw.c 7.2
SCCS-vsn: sys/netiso/clnp_stat.h 7.2
SCCS-vsn: sys/netiso/clnp_subr.c 7.4
SCCS-vsn: sys/netiso/clnp_timer.c 7.2
SCCS-vsn: sys/netiso/cons.h 7.2
SCCS-vsn: sys/netiso/cons_pcb.h 7.2
SCCS-vsn: sys/netiso/eonvar.h 7.2
SCCS-vsn: sys/netiso/esis.c 7.2
SCCS-vsn: sys/netiso/esis.h 7.2
SCCS-vsn: sys/netiso/if_clnp.h 7.2
SCCS-vsn: sys/netiso/if_cons.c 7.2
SCCS-vsn: sys/netiso/if_eon.c 7.2
SCCS-vsn: sys/netiso/if_lpb.c 7.2
SCCS-vsn: sys/netiso/iso.c 7.2
SCCS-vsn: sys/netiso/iso.h 7.2
SCCS-vsn: sys/netiso/iso_addrs.c 7.2
SCCS-vsn: sys/netiso/iso_chksum.c 7.2
SCCS-vsn: sys/netiso/iso_errno.h 7.2
SCCS-vsn: sys/netiso/iso_map.h 7.2
SCCS-vsn: sys/netiso/iso_pcb.c 7.2
SCCS-vsn: sys/netiso/iso_pcb.h 7.2
SCCS-vsn: sys/netiso/iso_proto.c 7.2
SCCS-vsn: sys/netiso/iso_snpac.c 7.2
SCCS-vsn: sys/netiso/iso_snpac.h 7.2
SCCS-vsn: sys/netiso/iso_usrreq.c 7.2
SCCS-vsn: sys/netiso/iso_var.h 7.2
SCCS-vsn: sys/netiso/tp.trans 7.2
SCCS-vsn: sys/netiso/tp_astring.c 7.2
SCCS-vsn: sys/netiso/tp_clnp.h 7.2
SCCS-vsn: sys/netiso/tp_cons.c 7.2
SCCS-vsn: sys/netiso/tp_emit.c 7.2
SCCS-vsn: sys/netiso/tp_inet.c 7.2
SCCS-vsn: sys/netiso/tp_input.c 7.2
SCCS-vsn: sys/netiso/tp_ip.h 7.2
SCCS-vsn: sys/netiso/tp_iso.c 7.2
SCCS-vsn: sys/netiso/tp_meas.c 7.2
SCCS-vsn: sys/netiso/tp_meas.h 7.2
SCCS-vsn: sys/netiso/tp_output.c 7.2
SCCS-vsn: sys/netiso/tp_param.h 7.2
SCCS-vsn: sys/netiso/tp_pcb.c 7.2
SCCS-vsn: sys/netiso/tp_pcb.h 7.2
SCCS-vsn: sys/netiso/tp_seq.h 7.2
SCCS-vsn: sys/netiso/tp_sizes.c 7.2
SCCS-vsn: sys/netiso/tp_stat.h 7.2
SCCS-vsn: sys/netiso/tp_subr.c 7.2
SCCS-vsn: sys/netiso/tp_subr2.c 7.2
SCCS-vsn: sys/netiso/tp_timer.c 7.2
SCCS-vsn: sys/netiso/tp_timer.h 7.2
SCCS-vsn: sys/netiso/tp_tpdu.h 7.2
SCCS-vsn: sys/netiso/tp_trace.c 7.2
SCCS-vsn: sys/netiso/tp_trace.h 7.2
SCCS-vsn: sys/netiso/tp_user.h 7.2
SCCS-vsn: sys/netiso/tp_usrreq.c 7.2

+++ usr/src/sys/netiso/tp_param.h
@@ -46,8 +46,8 @@ SOFTWARE.
 
 #define 	N_TPREF				100
 
-#define 	TP_SOCKBUFSIZE		4096
-#define 	TP0_SOCKBUFSIZE		512
+#define 	TP_SOCKBUFSIZE		((u_long)4096)
+#define 	TP0_SOCKBUFSIZE		((u_long)512)
 #define		MAX_TSAP_SEL_LEN	64
 
 /* maximum tpdu size we'll accept: */
@@ -284,7 +284,7 @@ bcopy((caddr_t)&(((struct tp_vbp *)(src))->tpv_val),(caddr_t)&(dst),sizeof(type)
 	P = (caddr_t)(DU) + (int)((DU)->tpdu_li);\
 	vbptr(P)->tpv_code = type;\
 	vbptr(P)->tpv_len = len;\
-	bcopy((caddr_t)&src, (caddr_t)&(vbptr(P)->tpv_val), (int)len);\
+	bcopy((caddr_t)&src, (caddr_t)&(vbptr(P)->tpv_val), (unsigned)len);\
 	DU->tpdu_li += len+2;/* 1 for code, 1 for length */\
 }
 /******************************************************
